standardize on long/double for statistics, and add various ValueStatistics for sampling purposes
git-svn-id: https://svn.terracotta.org/repo/forge/projects/statistics/trunk@37652 591de9ec-d6d5-471c-8416-afe4a41b277d

+++ src/main/java/org/terracotta/statistics/derived/EventRateSimpleMovingAverage.java
@@ -19,7 +19,7 @@ import static org.terracotta.statistics.Time.time;
  *
  * @author cdennis
  */
-public class EventRateSimpleMovingAverage implements EventObserver, ValueStatistic<Float> {
+public class EventRateSimpleMovingAverage implements EventObserver, ValueStatistic<Double> {
 
   private static final int PARTITION_COUNT = 10;
 
@@ -35,11 +35,11 @@ public class EventRateSimpleMovingAverage implements EventObserver, ValueStatist
   }
 
   @Override
-  public Float value() {
+  public Double value() {
     return rateUsingNanos();
   }
   
-  public Float rateUsingNanos() {
+  public Double rateUsingNanos() {
     long endTime = time();
     long startTime = endTime - windowSize;
     
@@ -65,10 +65,10 @@ public class EventRateSimpleMovingAverage implements EventObserver, ValueStatist
       }
     }
     
-    return ((float) count) / (endTime - actualStartTime);
+    return ((double) count) / (endTime - actualStartTime);
   }
   
-  public Float rate(TimeUnit base) {
+  public Double rate(TimeUnit base) {
     return rateUsingNanos() * base.toNanos(1);
   }
